🌱 [scanner] refactor: split lib/api.ts into endpoints and types modules (#21384)
